make refresh rate a float

+++ src/SFML/Window/VideoMode.cpp
@@ -32,7 +32,7 @@
 namespace sf
 {
 ////////////////////////////////////////////////////////////
-VideoMode::VideoMode(const Vector2u& modeSize, unsigned int modeRefreshRate, unsigned int modeBitsPerPixel) :
+VideoMode::VideoMode(const Vector2u& modeSize, float modeRefreshRate, unsigned int modeBitsPerPixel) :
 size(modeSize),
 refreshRate(modeRefreshRate),
 bitsPerPixel(modeBitsPerPixel)
